+++ packages/polaris-viz/src/components/VerticalBarChart/Chart.tsx
@@ -331,6 +331,7 @@ export function Chart({
           renderLegendContent={renderLegendContent}
           width={width}
           enableHideOverflow
+          dimensions={dimensions}
         />
       )}
     </ChartElements.Div>